Extract console code from main class to controller
